\name{cnolist}
\docType{data}
\alias{cnolist}
\title{CNOlist}
\description{CNOlist object containing the perturbation experimental data.}
\usage{cnolist}
\format{cnolist is the object which contains the data in the MIDAS file.}
\source{This object is generated from the dynamic-feeder example}
\keyword{datasets}
