\name{database}
\docType{data}
\alias{database}
\title{OmniPath PPI}
\description{Data-frame containing signed and directed interactions from Omnipath.}
\usage{database}
\format{database is the object which contains new interactions which can potentially be integrated.}
\source{This object is generated from the dynamic-feeder example}
\keyword{datasets}
