\name{feederObject}
\docType{data}
\alias{feederObject}
\title{Feeder Object}
\description{Object list as obtained from the buildFeederObjectDynamic() function.}
\usage{feederObject}
\format{feederObject is a list containing interactions suggested to be added in the PKN.}
\source{This object is generated from the dynamic-feeder example}
\keyword{datasets}
