\name{indices}
\docType{data}
\alias{indices}
\title{Mis-fit indices}
\description{Simulation data as obtained from the identifyMisfitIndices() function.}
\usage{indices}
\format{indices is a list of poorly predicted measurements.}
\source{This object is generated from the dynamic-feeder example}
\keyword{datasets}
