\name{integratedModel}
\docType{data}
\alias{integratedModel}
\title{Integrated Model}
\description{PKNlist object as obtained from the integrateLinks() function.}
\usage{integratedModel}
\format{integratedModel is the model we obtain after the integration of the new links.}
\source{This object is generated from the dynamic-feeder example}
\keyword{datasets}
