\name{model}
\docType{data}
\alias{model}
\title{Prior Knowledge Network}
\description{Model object from the dynamic-feeder example.}
\usage{model}
\format{model is an PKNlist with proteins as nodes and undirected links as physical protein interactions.}
\source{This object is generated from the dynamic-feeder example}
\keyword{datasets}
