\name{simData}
\docType{data}
\alias{simData}
\title{CNORode simuation data}
\description{Simulation data as obtained from the plotLBodeFitness() function.}
\usage{simData}
\format{simData is a list containing simulated values for a specific set of ode parameters.}
\source{This object is generated from the dynamic-feeder example}
\keyword{datasets}
