\name{UniprotIDdream}
\docType{data}
\alias{UniprotIDdream}
\title{Uniprot identifiers for proteins in DreamModel}
\description{This data object contains the Uniprot identifiers corresponding to DreamModel of CellNOptR package, in order to associat them with the corresponding nodes in the protein-protein interaction network (PPINigraph).}
\usage{UniprotIDdream}
\format{UniprotIDdream is a list where each element is a protien of the DreamModel and is associated with the respective Uniprot identifiers.}
\source{This data object is manually derived from the Uniprot database.}
\references{
\enumerate{
\item{F. Eduati, J. De Las Rivas, B. Di Camillo, G. Toffolo, J. Saez-Rodriguez. Integrating literature-constrained and data-driven inference of signalling networks. Bioinformatics, 28(18):2311-2317, 2012.}
\item{J. Saez-Rodriguez, L. G. Alexopoulos, J. Epperlein, R. Samaga, D. A.
Lauffenburger, S. Klamt and P. K. Sorger. Discrete logic modeling as a means to
link protein signaling networks with functional analysis of mammalian signal
transduction, Molecular Systems Biology, 5:331, 2009.}
}
}
\keyword{datasets}
